8.18 cont.: add logout button
Let's also add a button which enables a logged in user to log out. The
button's onClick handler sets the _token_ state to null, removes the
token from local storage and resets the cache of the Apollo
client. The last step is
[important](https://www.apollographql.com/docs/react/networking/authentication/#reset-store-on-logout),
because some queries might have fetched data to cache, which only
logged in users should have access to.

We can reset the cache using the
[resetStore](https://www.apollographql.com/docs/react/v3.0-beta/api/core/ApolloClient/#ApolloClient.resetStore)
method of an Apollo _client_ object.

The client can be accessed with the
[useApolloClient](https://www.apollographql.com/docs/react/api/react-hooks/#useapolloclient)
hook.