Finish 8.18 Log in
Adding new books and changing the birth year of an author do not work
because they require user to be logged in.

Implement login functionality and fix the mutations.

It is not necessary yet to handle validation errors.

===========

After the backend changes, creating new persons requires that a valid
user token is sent with the request. In order to send the token, we
have to change the way we define the _ApolloClient_-object in
_index.js_ a little.

The link parameter given to the _client_-object defines how apollo
connects to the server. Here the normal
[httpLink](https://www.apollographql.com/docs/link/links/http.htm)
connection is modified so that the request's _authorization_
[header](https://www.apollographql.com/docs/react/networking/authentication/#header)
contains the token if one has been saved to the localStorage.

We also need to install the library required by this modification

```js
npm install --save apollo-link-context
```