8.21 cont.: finish allBooks query with optional arguments
[$in](https://docs.mongodb.com/manual/reference/operator/query/in/)

==========

**Example**

Complete the _allPersons_ resolver so it takes the optional parameter
_phone_ into account:

```js
Query: {
  // ..
  allPersons: (root, args) => {
    if (!args.phone) {
      return Person.find({})
    }

    return Person.find({ phone: { $exists: args.phone === 'YES'  }})
  },
},
```

So if the query has not been given a parameter _phone_, all persons
are returned. If the parameter has the value _YES_, the result of the
query

```js
Person.find({ phone: { $exists: true }})
```

is returned, so the objects in which the field _phone_ has a value. If
the parameter has the value _NO_, the query returns the objects in
which the _phone_ field has no value:

```js
Person.find({ phone: { $exists: false }})
```