Finish 8.23: Subscriptions - server
Do a backend implementation for subscription bookAdded, which returns
the details of all new books to its subscribers.

========

The schema changes like so:

```js
type Subscription {
  bookAdded: Book!
}
```

So when a new book is added, all of its details are sent to all subscribers.

The subscription _bookAdded_ needs a resolver. The _addBook_
resolver also has to be modified so that it sends a notification to
subscribers.

With subscriptions, the communication happens using the
[publish-subscribe](https://en.wikipedia.org/wiki/Publish%E2%80%93subscribe_pattern)
principle utilizing an object using a
[PubSub](https://www.apollographql.com/docs/graphql-subscriptions/setup/#setup)
interface. Adding a new book _publishes_ a notification about the
operation to all subscribers with PubSub's method _publish_.

_bookAdded_ subscriptions resolver registers all of the subscribers
by returning them a suitable [iterator
object](https://www.apollographql.com/docs/graphql-subscriptions/subscriptions-to-schema/).

Let's do the following changes to the code which starts the server

```js
// ...

server.listen().then(({ url, subscriptionsUrl }) => {  console.log(`Server ready at ${url}`)
  console.log(`Subscriptions ready at ${subscriptionsUrl}`)})
```

We see, that the server listens for subscriptions in the address
_ws://localhost:4000/graphql_

```js
Server ready at http://localhost:4000/
Subscriptions ready at ws://localhost:4000/graphql
```

No other changes to the server are needed.

It's possible to test the subscriptions with the GraphQL playground
like this:

```gql
subscription {
  bookAdded {
    title
    published
    author {
      name
    }
    genres
  }
}
```

When you press "play" on a subscription, the playground waits for
notifications from the subscription.