Finish 8.24: Subscriptions - client, part 1
Start using subscriptions in the client, and subscribe to
_bookAdded_. When new books are added, notify the user. Any method
works. For example, you can use the
[window.alert](https://developer.mozilla.org/en-US/docs/Web/API/Window/alert)
function.

========

In order to use subscriptions in our React application, we have to do
some changes, especially on its
[configuration](https://www.apollographql.com/docs/react/v3.0-beta/data/subscriptions/).

The configuration in _index.js_ has to be modified.

For this to work, we have to install some dependencies:

```js
npm install --save @apollo/link-ws subscriptions-transport-ws
```

The new configuration is due to the fact that the application must
have an HTTP connection as well as a WebSocket connection to the
GraphQL server.

```js
const wsLink = new WebSocketLink({
  uri: `ws://localhost:4000/graphql`,
  options: { reconnect: true }
})

const httpLink = createHttpLink({
  uri: 'http://localhost:4000',
})
```

The subscriptions are done using the
[useSubscription](https://www.apollographql.com/docs/react/v3.0-beta/api/react/hooks/#usesubscription)
hook function.

When a new person is added, the server sends a notification to the
client, and the callback-function defined in the _onSubscriptionData_
attribute is called and given the details of the new person as
parameters.