Finish 8.25: Subscriptions - client, part 2
Keep the application's view updated when the server notifies about new
books. You can test your implementation by opening the app in two
browser tabs and adding a new book in one tab. Adding the new book
should update the view in both tabs.

===============

Let's extend our solution so that when the details of a new person are
received, the person is added to the Apollo cache, so it is rendered
to the screen immediately.

However, we have to keep in mind that when our application creates a
new person, it should not be added to the cache twice.

The function _updateCacheWith_ can also be used in _PersonForm_ for
the cache update:

```language-js
const PersonForm = ({ setError, updateCacheWith }) => {
  // ...

  const [ createPerson ] = useMutation(CREATE_PERSON, {
    onError: (error) => {
      setError(error.graphQLErrors[0].message)
    },
    update: (store, response) => {
      updateCacheWith(response.data.addPerson)
    }
  })

  // ..
}
```